cleanstore/requirements.txt
africastalking==1.2.7
asgiref==3.7.2
blessed==1.20.0
cement==2.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.3
compose==1.5.0
contextlib2==21.6.0
cryptography==42.0.5
Cython==3.0.8
defusedxml==0.7.1
dj-database-url==2.1.0
Django==5.0.2
django-allauth==0.61.1
django-filter==24.1
django-storages==1.14.2
djangorestframework==3.14.0
docker==7.0.0
gunicorn==21.2.0
idna==3.6
jmespath==1.0.1
oauthlib==3.2.2
packaging==23.2
pathspec==0.10.1
phonenumbers==8.13.31
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.10.0
schema==0.7.5
semantic-version==2.8.5
six==1.16.0
sqlparse==0.4.4
termcolor==1.1.0
typing_extensions==4.10.0
urllib3==1.26.18
wcwidth==0.1.9
whitenoise==6.6.0
